[Behat][Admin] Introduce shared FormElement
