+++ src/Sylius/Behat/Element/Admin/Crud/FormElementInterface.php
@@ -1,0 +1,22 @@
+<?php
+
+/*
+ * This file is part of the Sylius package.
+ *
+ * (c) Sylius Sp. z o.o.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Sylius\Behat\Element\Admin\Crud;
+
+use Behat\Mink\Exception\ElementNotFoundException;
+
+interface FormElementInterface
+{
+    /** @throws ElementNotFoundException */
+    public function getValidationMessage(string $element): string;
+}
