+++ src/Sylius/Behat/Resources/config/services/elements/admin.xml
@@ -16,6 +16,8 @@
            xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd"
 >
     <services>
+        <service id="sylius.behat.element.admin.crud.form" class="Sylius\Behat\Element\Admin\Crud\FormElement" parent="sylius.behat.element" />
+
         <service id="sylius.behat.element.admin.account.reset" class="Sylius\Behat\Element\Admin\Account\ResetElement" parent="sylius.behat.element" />
 
         <service id="sylius.behat.element.admin.channel.shipping_address_in_checkout_required" class="Sylius\Behat\Element\Admin\Channel\ShippingAddressInCheckoutRequiredElement" parent="sylius.behat.element" />
@@ -55,7 +57,7 @@
         <service
             id="Sylius\Behat\Element\Admin\Promotion\FormElementInterface"
             class="Sylius\Behat\Element\Admin\Promotion\FormElement"
-            parent="sylius.behat.element"
+            parent="sylius.behat.element.admin.crud.form"
         >
             <argument type="service" id="Sylius\Behat\Service\Helper\AutocompleteHelperInterface" />
         </service>
